Succeed conv split with resnet8

+++ tvm_practice/models/resnet8_cifar.py
@@ -45,7 +45,7 @@ def getModel_(input_shape):
   y = imcflow_min_max_quantize(y, relay.var("quant_min_1", shape=(), dtype="int16"), relay.var("quant_max_1", shape=(), dtype="int16"), axis=1, out_dtype="int4")
   y = imcflow_qconv2d(
     y,
-    relay.var("weight2_1", shape=(16,16,3,3), dtype="int4"),
+    relay.var("weight2_1", shape=(16,16,3,3), dtype="int8"),
     in_channels=16,
     channels=16,
     kernel_size=(3, 3),
@@ -56,7 +56,7 @@ def getModel_(input_shape):
   y = imcflow_min_max_quantize(y, relay.var("quant_min_2", shape=(), dtype="int16"), relay.var("quant_max_2", shape=(), dtype="int16"), axis=1, out_dtype="int4")
   y = imcflow_qconv2d(
     y,
-    relay.var("weight2_2", shape=(16,16,3,3), dtype="int4"),
+    relay.var("weight2_2", shape=(16,16,3,3), dtype="int8"),
     in_channels=16,
     channels=16,
     kernel_size=(3, 3),
@@ -71,7 +71,7 @@ def getModel_(input_shape):
   y = imcflow_min_max_quantize(y, relay.var("quant_min_3", shape=(), dtype="int16"), relay.var("quant_max_3", shape=(), dtype="int16"), axis=1, out_dtype="int4")
   y = imcflow_qconv2d(
     y,
-    relay.var("weight3_1", shape=(32,16,3,3), dtype="int4"),
+    relay.var("weight3_1", shape=(32,16,3,3), dtype="int8"),
     in_channels=16,
     channels=32,
     kernel_size=(3, 3),
@@ -83,7 +83,7 @@ def getModel_(input_shape):
   y = imcflow_min_max_quantize(y, relay.var("quant_min_4", shape=(), dtype="int16"), relay.var("quant_max_4", shape=(), dtype="int16"), axis=1, out_dtype="int4")
   y = imcflow_qconv2d(
     y,
-    relay.var("weight3_2", shape=(32,32,3,3), dtype="int4"),
+    relay.var("weight3_2", shape=(32,32,3,3), dtype="int8"),
     in_channels=32,
     channels=32,
     kernel_size=(3, 3),
@@ -95,7 +95,7 @@ def getModel_(input_shape):
   y_residual = imcflow_min_max_quantize(residual, relay.var("quant_min_4_2", shape=(), dtype="int16"), relay.var("quant_max_4_2", shape=(), dtype="int16"), axis=1, out_dtype="int4")
   y_residual = imcflow_qconv2d(
     y_residual,
-    relay.var("weight3_0", shape=(32,16,1,1), dtype="int4"),
+    relay.var("weight3_0", shape=(32,16,1,1), dtype="int8"),
     in_channels=16,
     channels=32,
     kernel_size=(1, 1),
@@ -110,7 +110,7 @@ def getModel_(input_shape):
   y = imcflow_min_max_quantize(y, relay.var("quant_min_5", shape=(), dtype="int16"), relay.var("quant_max_5", shape=(), dtype="int16"), axis=1, out_dtype="int4")
   y = imcflow_qconv2d(
     y,
-    relay.var("weight4_1", shape=(64,32,3,3), dtype="int4"),
+    relay.var("weight4_1", shape=(64,32,3,3), dtype="int8"),
     in_channels=32,
     channels=64,
     kernel_size=(3, 3),
@@ -122,7 +122,7 @@ def getModel_(input_shape):
   y = imcflow_min_max_quantize(y, relay.var("quant_min_6", shape=(), dtype="int16"), relay.var("quant_max_6", shape=(), dtype="int16"), axis=1, out_dtype="int4")
   y = imcflow_qconv2d(
     y,
-    relay.var("weight4_2", shape=(64,64,3,3), dtype="int4"),
+    relay.var("weight4_2", shape=(64,64,3,3), dtype="int8"),
     in_channels=64,
     channels=64,
     kernel_size=(3, 3),
@@ -134,7 +134,7 @@ def getModel_(input_shape):
   y_residual = imcflow_min_max_quantize(residual, relay.var("quant_min_6_2", shape=(), dtype="int16"), relay.var("quant_max_6_2", shape=(), dtype="int16"), axis=1, out_dtype="int4")
   y_residual = imcflow_qconv2d(
     y_residual,
-    relay.var("weight4_0", shape=(64,32,1,1), dtype="int4"),
+    relay.var("weight4_0", shape=(64,32,1,1), dtype="int8"),
     in_channels=32,
     channels=64,
     kernel_size=(1, 1),
